Rename "Half of screen" mode to "Fit on screen"
The goal of this zoom mode is to fit the image (along with An Image
Viewer's UI) onto the screen so that nothing reaches out. If an image is
small enough, then zoom mode 100% is simply used, but for larger images,
An Image Viewer currently (arbitrarily) adds 200px to the image's height
to make up for its toolbar and window decoration.

If the height after this calculation is larger than the screen height,
"Fit on screen" is being used which will shrink the image slightly.